requirements.txt
discord.py==2.2.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
jsons==1.6.3
nest_asyncio==1.5.6
pymongo==4.4.0b0
